refactor(utils): move adapter utils
