+++ lua/codecompanion/adapters/http/gemini.lua
@@ -1,9 +1,9 @@
 ---Source: https://ai.google.dev/gemini-api/docs
 
-local adapter_utils = require("codecompanion.utils.adapters")
+local adapter_utils = require("codecompanion.adapters.utils")
 local log = require("codecompanion.utils.log")
 local tags = require("codecompanion.interactions.shared.tags")
-local transform = require("codecompanion.utils.tool_transformers")
+local transform = require("codecompanion.adapters.utils.tool_transformers")
 
 ---Extract the first complete JSON object from a potentially concatenated string
 ---Workaround for Gemini bug where multiple JSON objects get concatenated
