+++ lua/codecompanion/adapters/http/openai_compatible.lua
@@ -5,7 +5,7 @@ to CodeCompanion via an adapter. Send any questions or queries to the discussion
 --]]
 
 local Curl = require("plenary.curl")
-local adapter_utils = require("codecompanion.utils.adapters")
+local adapter_utils = require("codecompanion.adapters.utils")
 local config = require("codecompanion.config")
 local log = require("codecompanion.utils.log")
 local openai = require("codecompanion.adapters.http.openai")
