+++ lua/codecompanion/adapters/http/openrouter.lua
@@ -1,5 +1,5 @@
 local Curl = require("plenary.curl")
-local adapter_utils = require("codecompanion.utils.adapters")
+local adapter_utils = require("codecompanion.adapters.utils")
 local config = require("codecompanion.config")
 local log = require("codecompanion.utils.log")
 local openai = require("codecompanion.adapters.http.openai")
