+++ lua/codecompanion/adapters/shared.lua
@@ -1,4 +1,4 @@
-local adapter_utils = require("codecompanion.utils.adapters")
+local adapter_utils = require("codecompanion.adapters.utils")
 local log = require("codecompanion.utils.log")
 
 local M = {}
