+++ lua/codecompanion/adapters/utils/init.lua
