+++ lua/codecompanion/adapters/utils/tool_transformers.lua
