+++ tests/adapters/http/test_ollama.lua
@@ -348,7 +348,7 @@ end
 T["Ollama adapter"]["OLLAMA_HOST"] = new_set()
 
 T["Ollama adapter"]["OLLAMA_HOST"]["uses OLLAMA_HOST when set"] = function()
-  local adapter_utils = require("codecompanion.utils.adapters")
+  local adapter_utils = require("codecompanion.adapters.utils")
   vim.env.OLLAMA_HOST = "http://192.168.1.100:11434"
   adapter_utils.get_env_vars(adapter)
   h.eq("http://192.168.1.100:11434", adapter.env_replaced.url)
@@ -356,7 +356,7 @@ T["Ollama adapter"]["OLLAMA_HOST"]["uses OLLAMA_HOST when set"] = function()
 end
 
 T["Ollama adapter"]["OLLAMA_HOST"]["fallback to localhost when OLLAMA_HOST is not set"] = function()
-  local adapter_utils = require("codecompanion.utils.adapters")
+  local adapter_utils = require("codecompanion.adapters.utils")
   vim.env.OLLAMA_HOST = nil
   adapter_utils.get_env_vars(adapter)
   h.eq("http://localhost:11434", adapter.env_replaced.url)
